% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_phosphosites.R
\name{build_fastas}
\alias{build_fastas}
\title{Create a FASTA file with sequence fragments}
\usage{
build_fastas(data, path, name_col, seq_col, name_pattern = NULL)
}
\arguments{
\item{data}{Data frame in long format, containing unique names and sequence windows.}

\item{path}{Path to write the file to, including file name and extension.}

\item{name_col}{Name of the column that contains unique fasta IDs.
If missing, uses name_pattern to construct unique IDs.}

\item{seq_col}{Name of the column that contains the sequence windows.}

\item{name_pattern}{Optional, only used if \code{name_col} is missing.
Pattern for use in \code{\link[glue:glue]{glue::glue()}}, used to construct a new unique name column
based on current columns.}
}
\value{
Returns path of the output FASTA, invisibly.
}
\description{
From (potential) sites and their surrounding amino acids, create a FASTA-conforming file.
Requires a column of unique names for each site.
}
\examples{
kinsub_path <- system.file('extdata', 'Kinase_Substrate_Dataset_head', package = 'phosphocie')
kinsub <- phosphocie:::read_kinsub(kinsub_path)
tmp <- tempfile()

build_fastas(kinsub, tmp, name_col = 'unique_id', seq_col = 'fragment_15')

build_fastas(kinsub, tmp, seq_col = 'fragment_15', name_pattern = '{acc_id}|{gene}|{substrate}|{residue}{position}')
}
