% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_phosphosites.R
\name{filter_phosphosites}
\alias{filter_phosphosites}
\title{Filter phosphosite data to human, uniprot-conformant data}
\usage{
filter_phosphosites(data, organism_col = "organism", acc_id_col = "acc_id")
}
\arguments{
\item{data}{Dataset from \code{\link[=read_phosphosite]{read_phosphosite()}} or \code{\link[=read_kinsub]{read_kinsub()}}}

\item{organism_col}{Optional: name of data column containing organism data.}

\item{acc_id_col}{Optional: name of data column containing accession id data.}
}
\value{
The filtered data frame.
}
\description{
Filters out rows from non-human organisms or with non-uniprot-conformant accession ids.
}
\keyword{internal}
