% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucie_transformations.R
\name{fit_transform}
\alias{fit_transform}
\title{Calculate UCIE geometric transformation values}
\usage{
fit_transform(dataset, rownames_col = NULL, center = TRUE)
}
\arguments{
\item{dataset}{A data frame or matrix of 3D values.}

\item{rownames_col}{Optional: name of column to ignore during calculations.}

\item{center}{A boolean indicating whether data should be centered during the fitting
process. Improves performance, but requires \code{center = TRUE} in \code{\link[=transform_data]{transform_data()}} too.}
}
\value{
Returns a numeric vector with a scaling value (S), 3 rotation values
(RotL, Rota, Rotb), and 3 translation values (TrL, Tra, Trb) for use in \code{\link[=transform_data]{transform_data()}}.
}
\description{
Scales, rotates and transforms the data to best fit in the CIELAB colourspace.
Wrapper around ucie:::FitColorsFunction()
}
\details{
\code{dataset} can be a 2D or 3D matrix or data frame, with optional
rownames in the first column or a \code{rownames_col}-defined column.
}
