% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_phosphosites.R
\name{get_middle_fragment}
\alias{get_middle_fragment}
\title{Extract middle fragment from each string}
\usage{
get_middle_fragment(string_vec, size)
}
\arguments{
\item{string_vec}{A character vector of strings, all at least size \code{size}.}

\item{size}{The size of the returned middle fragments, as an uneven integer.}
}
\value{
An all-uppercase character vector, of the same length as the \code{string_vec} input.
}
\description{
Returns an uppercase slice of size \code{size} from each string supplied, centered
on the middle position of the string. Useful to slice out middles from
sequence fragments.
}
