% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucie_transformations.R
\name{kinase2cielab}
\alias{kinase2cielab}
\title{Get colours from fitted kinase data}
\usage{
kinase2cielab(
  dataset,
  transform_vals,
  LAB_coordinates = FALSE,
  rownames_col = NULL
)
}
\arguments{
\item{dataset}{A data frame or matrix of 3D values.}

\item{transform_vals}{A numeric vector with 1 scaling, 3 rotation, and 3 translation values,
as calculated by \code{\link[=ucie_transformations]{ucie_transformations()}}.}

\item{LAB_coordinates}{Optional: boolean, whether to return CIELAB coordinates instead of RGB codes.}

\item{rownames_col}{Optional: name of column to ignore during calculations.}
}
\value{

}
\description{
Get colours from fitted kinase data
}
\details{
\code{dataset} can be a 2D or 3D matrix or data frame, with optional
rownames in the first column or a \code{rownames_col}-defined column.
}
