% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucie_transformations.R
\name{kinase2refcielab}
\alias{kinase2refcielab}
\title{Get kinase score colours based on reference colourspace}
\usage{
kinase2refcielab(data, k = 10, rownames_col = NULL, fix = TRUE)
}
\arguments{
\item{data}{A data frame or matrix, with the same columns as the reference
kinase scoring dataset (NetPhorest, 61 kinases), and optionally a rownames
column (see \code{rownames_col})}

\item{k}{Number of neighbours to consider.}

\item{rownames_col}{Optional: name of a column containing rownames, set as
true rownames before continuing. Allows usage of tidy data in this matrix-based function.}

\item{fix}{Optional: boolean, whether to force points outside the colour space inside (TRUE)
or return NA (FALSE). Default is TRUE.}
}
\value{
A data frame with
}
\description{
Get kinase score colours based on reference colourspace
}
