% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosphocie.R
\docType{package}
\name{phosphocie}
\alias{phosphocie}
\title{phosphocie: A package to visualise phosphosite kinase profiles}
\description{
The phosphocie package builds on ucie to visualise kinase preferences of
phosphosite predictions in a continuous, perceptually uniform colourspace.
}
