% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_phosphosites.R
\name{read_kinsub}
\alias{read_kinsub}
\title{Read in PhosphoSitePlus kinase substrate data}
\usage{
read_kinsub(path)
}
\arguments{
\item{path}{The path to the unzipped kinase substrate file.}
}
\description{
Wrapper around \code{\link[readr:read_delim]{readr::read_tsv()}} that cleans up the default kinase-substrate
data to have nicer columns, cleaner values, and a prepared unique site ID.
}
\examples{
kinsub_path <- system.file('extdata', 'Kinase_Substrate_Dataset_head', package = 'phosphocie')
kinsub <- read_kinsub(kinsub_path)

}
\keyword{internal}
