% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_kinases.R
\name{read_ksp}
\alias{read_ksp}
\title{Read in KSP kinase prediction output}
\usage{
read_ksp(path)
}
\arguments{
\item{path}{Path to the \verb{*.*_kspoutput} file}
}
\value{
A matrix with of sites x (detected) kinases.
}
\description{
Read in KSP kinase prediction output
}
