% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_phosphosites.R
\name{read_phosphosite}
\alias{read_phosphosite}
\title{Read in PhosphoSitePlus phosphorylation site data}
\usage{
read_phosphosite(path)
}
\arguments{
\item{path}{The path to the unzipped kinase substrate file.}
}
\description{
Wrapper around \code{\link[readr:read_delim]{readr::read_tsv()}} that cleans up the default phosphorylation
site data to have nicer columns, cleaner values, and a prepared unique site ID.
}
\examples{
phosphosite_path <- system.file('extdata', 'Phosphorylation_Site_Dataset_head', package = 'phosphocie')
phosphosite <- read_phosphosite(phosphosite_path)

}
\keyword{internal}
