% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_umap.R
\docType{data}
\name{ref_umap}
\alias{ref_umap}
\title{\code{uwot::umap} pre-trained PhosphoSitePlus reference model}
\format{
A list, containing the elements of an UMAP model:
\describe{
\item{embedding}{3D coordinates, as a (239419, 3) matrix}
\item{...}{parameters of the model}
}
}
\source{
\url{https://www.phosphosite.org/staticDownloads}
}
\usage{
ref_umap
}
\description{
An object for use with \code{\link[uwot:umap_transform]{uwot::umap_transform()}}, containing the 3D
representation of the PhosphoSitePlus phosphorylation sites dataset,
which contains almost 240000 phosphorylation sites. Originally trained
with spread = 5, using the code in \code{data_raw/create_reference_colourspace.R}.
}
\keyword{datasets}
