% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_phosphosites.R
\name{retrieve_fastas}
\alias{retrieve_fastas}
\title{Retrieve full protein sequences from UniProt}
\usage{
retrieve_fastas(uniprot_acc, path)
}
\arguments{
\item{uniprot_acc}{Character vector of UniProt accession IDs. Supports isoform IDs like \code{Q02297-6}.}

\item{path}{Path to write the file to, including file name and extension.}
}
\value{
Returns path of the output FASTA, invisibly.
}
\description{
From a vector of uniprot accession IDs, retrieves FASTAs for each protein from
the \href{https://www.ebi.ac.uk/proteins/api/doc/index.html}{EMBL Proteins API} and
concatenates these into one massive FASTA file.
Can serve as NetPhorest input, but \code{\link[=build_fastas]{build_fastas()}} is the recommended alternative
to predict for known sites rather than entire proteins.
}
\details{
\itemize{
\item Filters out duplicated IDs and invalid IDs.
\item Requests in batches of 100, which is the max for the API.
\item If any batch fails, the process is stopped and the file is left as is, containing all
successful batches up to then.
\item Between each batch request, waits for 0.75 seconds to stay below rate limits.
}
}
\examples{
kinsub_path <- system.file('extdata', 'Kinase_Substrate_Dataset_head', package = 'phosphocie')
kinsub <- phosphocie:::read_kinsub(kinsub_path)
tmp <- tempfile()

\dontrun{
  retrieve_fastas(kinsub$acc_id, tmp)
}
}
