% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucie_transformations.R
\name{transform_data}
\alias{transform_data}
\title{Get colours from fitted kinase data}
\usage{
transform_data(
  dataset,
  transform_vals,
  LAB_coordinates = FALSE,
  rownames_col = NULL,
  fix = TRUE,
  center = TRUE
)
}
\arguments{
\item{dataset}{A data frame or matrix of 3D values.}

\item{transform_vals}{A numeric vector with 1 scaling, 3 rotation, and 3 translation values,
as calculated by \code{\link[=fit_transform]{fit_transform()}}.}

\item{LAB_coordinates}{Optional: boolean, whether to return CIELAB coordinates instead of RGB codes.}

\item{rownames_col}{Optional: name of column to ignore during calculations.}

\item{fix}{Optional: boolean, whether to force points outside the colour space inside (TRUE)
or return NA (FALSE). Default is TRUE.}
}
\value{
Returns a data frame with names and hex codes in \code{name} and \code{colour}
(or \code{name}/\code{L}/\code{a}/\code{b} if \code{LAB_coordinates = TRUE}).
}
\description{
Get colours from fitted kinase data
}
\details{
\code{dataset} can be a 2D or 3D matrix or data frame, with optional
rownames in the first column or a \code{rownames_col}-defined column.
}
