% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucie_transformations.R
\name{ucie_transformations}
\alias{ucie_transformations}
\title{Calculate UCIE geometric transformation values}
\usage{
ucie_transformations(dataset, rownames_col = NULL)
}
\arguments{
\item{dataset}{A data frame or matrix of 3D values.}

\item{rownames_col}{Optional: name of column to ignore during calculations.}
}
\value{
Returns a numeric vector with a scaling value (S), 3 rotation values
(RotL, Rota, Rotb), and 3 translation values (TrL, Tra, Trb) for use in \code{\link[=kinase2cielab]{kinase2cielab()}}.
}
\description{
Scales, rotates and transforms the data to best fit in the CIELAB colourspace.
Wrapper around ucie:::FitColorsFunction()
}
\details{
\code{dataset} can be a 2D or 3D matrix or data frame, with optional
rownames in the first column or a \code{rownames_col}-defined column.
}
