% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_phosphosites.R
\name{write_ksp_input}
\alias{write_ksp_input}
\title{Create an input file for KSP}
\usage{
write_ksp_input(data, path, name_col = "gene", seq_col = "fragment_15")
}
\arguments{
\item{data}{Data frame in long format, containing gene names and sequence windows.}

\item{path}{Path to write the file to, including file name and extension.}

\item{name_col}{Name of the column that contains gene names.}

\item{seq_col}{Name of the column that contains sequence windows.}
}
\value{
Returns path that file was written to, invisibly.
}
\description{
Creates an input file to use for prediction with \href{https://doi.org/10.1186/s12864-020-06895-2}{KSP}.
Wrapper around \code{\link[readr:write_delim]{readr::write_tsv()}}.
}
